bot_hub/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async support
httpx==0.25.2
aiofiles==23.2.1

# Redis for rate limiting
redis==5.0.1

# Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
